LayoutTests/http/tests/preload/onerror_event-expected.txt

CONSOLE MESSAGE: line 25: <link rel=preload> must have a valid `as` value
This test makes sure that link preload triggers error events for non-existing resources.

PASS Makes sure that preloaded resources trigger onerror 
